src/web-service/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
jinja2
python-multipart
oryks-google-drive
requests
stripe
pyjwt
authlib
itsdangerous
pwdlib[argon2]
redis
celery[redis]
pygments